+++ src/Commands/Clean.php
@@ -37,19 +37,22 @@ class Clean extends Command
      */
     public function handle()
     {
-        $this->greetings();
-        $project_name = env('PROJECT_NAME');
-        if(!empty($project_name)){
-            exec('sudo rm -rf /var/www/html/'.env('PROJECT_NAME').'/bootstrap/cache/*');
-            $this->call('cache:clear');
-            $this->call('route:clear');
-            $this->call('view:clear');
-            $this->call('config:clear');
-            $this->call('clear-compiled');
-            exec('composer dump-autoload -o');
-        }else{
-            $this->error('=> Please, set PROJECT_NAME variable on .env file');
+        $project_name = env('PROJECT_DIRECTORY_NAME');
+        if (empty($project_name)) {
+            $this->error('=> need set PROJECT_DIRECTORY_NAME variable on .env file');
+            return;
         }
+
+        $this->greetings();
+        if (is_dir('/var/www/html/' . $project_name . '/bootstrap/cache/*')) {
+            exec('sudo rm -rf /var/www/html/' . $project_name . '/bootstrap/cache/*');
+        }
+        $this->call('cache:clear');
+        $this->call('route:clear');
+        $this->call('view:clear');
+        $this->call('config:clear');
+        $this->call('clear-compiled');
+        exec('composer dump-autoload -o');
         $this->farewell();
     }